requirements.txt
asgiref==3.5.2
bleach==5.0.1
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-static==0.0.6
Django==4.0.7
django-cloudinary-storage==0.3.0
django-colorfield==0.7.2
django-summernote==0.8.20.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
mssql-django==1.1.3
Pillow==9.2.0
pyodbc==4.0.34
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
wincertstore==0.2
aiohttp==3.8.4